rename {TableReader => table_reader}.hpp
